% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_tw_gender}
\alias{cpp_tw_gender}
\title{Matches names with genders}
\usage{
cpp_tw_gender(x, male, female)
}
\arguments{
\item{x}{Character vector with names}

\item{male}{Character vector with male names}

\item{female}{Character vector with female names}
}
\value{
Numeric vector with -1, 0, and 1 as 'unidentified', 'female', 'male'
}
\description{
Internal use, please refer to \code{\link{tw_gender}} instead.
}

