% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_processing.R
\name{jaccard_coef}
\alias{jaccard_coef}
\alias{Words}
\alias{similarities}
\alias{jaccard_coef.list}
\alias{jaccard_coef.character}
\title{Jaccard coefficient}
\usage{
jaccard_coef(x, ...)

\method{jaccard_coef}{list}(x, max.size = 1000, dist = FALSE, ...)

\method{jaccard_coef}{character}(x, max.size = 1000,
  stopwds = unique(c(tm::stopwords(), letters)), ignore.case = TRUE,
  dist = FALSE, ...)
}
\arguments{
\item{x}{Character vector with the phrases (tweets) to be analyzed.}

\item{...}{Further arguments to be passed to the method.}

\item{max.size}{Max number of words to analyze.}

\item{dist}{When true computes one minus Jaccard coef.}

\item{stopwds}{Character vector of stopwords.}

\item{ignore.case}{When true converts all to lower.}
}
\value{
A list including a lower triangular \code{dgCMatrix} matrix.
}
\description{
Calculate the Jaccard (similarity) coefficient between words.
}
\details{
The Jaccard index is used as a measure of similarity between two
elements. In particular for a given pair of elements \eqn{x,y} it is calculated as
\deqn{J(S,T) = \frac{|S\cap T|}{|S\cup T|}}{J(S,T) = |S intersection T|/|S U T|}
Where \eqn{S} is the set of groups where \eqn{x} is present and 
\eqn{T} is the set of groups where \eqn{y}. The resulting value is defined
between 0 and 1, where 0 corresponds to no similarity at all (the elements
don't have a group in common) and 1 represents perfect similarity (both elements
are present in the same groups).
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Method Processes a list of character vectors such as
the one obtained from \code{\link{tw_extract}}

\item \code{character}: Computes the coef from a vector of characters
(splits the text)
}}

\references{
Conover, M., Ratkiewicz, J., & Francisco, M. (2011). "Political polarization
on twitter". Icwsm, 133(26), 89–96. \url{http://doi.org/10.1021/ja202932e}
}
