% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{male_names_en_db}
\alias{male_names_en_db}
\title{English male names}
\format{A character vector with 51,212 elements}
\source{
OpenGenderTracking project \url{https://github.com/OpenGenderTracking/globalnamedata}
}
\description{
English male names
}

