% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{names_female_en}
\alias{names_female_en}
\title{English female names}
\format{A character vector with 80,148 elements}
\source{
OpenGenderTracking project \url{https://github.com/OpenGenderTracking/globalnamedata}
}
\description{
English female names
}

