% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{names_female_es}
\alias{names_female_es}
\title{Spanish female names}
\format{A character vector with 24,494 elements}
\source{
\url{https://github.com/apalancat/spanish-names}
}
\description{
Spanish female names
}

