% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{names_male_es}
\alias{names_male_es}
\title{Spanish male names}
\format{A character vector with 25,442 elements}
\source{
\url{https://github.com/apalancat/spanish-names}
}
\description{
Spanish male names
}

