% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.tw_Class_graph}
\alias{plot.tw_Class_graph}
\title{Plots a network using D3}
\usage{
\method{plot}{tw_Class_graph}(x, y = NULL, nodelabel = "name",
  opacity = 0.9, opacityNoHover = 0.5, legend = TRUE, ...)
}
\arguments{
\item{x}{A tw_Class_graph object (see tw_network)}

\item{y}{Not used}

\item{...}{Further arguments to pass to forceNetwork}
}
\value{
A D3 graph
}
\description{
Plots a network using D3
}

