% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.tw_Class_table}
\alias{plot.tw_Class_table}
\title{Plot a table}
\usage{
\method{plot}{tw_Class_table}(x, y = NULL, nentries = 20,
  caption = "Most popular hashtags", options = list(pageLength = 5), ...)
}
\description{
Plot a table
}

