% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.tw_Class_ts}
\alias{plot.tw_Class_ts}
\title{Plot time series}
\usage{
\method{plot}{tw_Class_ts}(x, rangeSelector = TRUE,
  main = "Number of daily tweets", xlab = NULL, ylab = NULL,
  group = NULL, width = 600, height = 400, ...)
}
\arguments{
\item{x}{a tw_Class_ts class object}

\item{rangeSelector}{Wheather or not to include a time range selector}

\item{main}{See dygraph}

\item{xlab}{See dygraph}

\item{ylab}{See dygraph}

\item{group}{See dygraph}

\item{height}{See dygraph}

\item{...}{Further arguments to be passed to dygraph}

\item{with}{See dygraph}
}
\description{
Plot time series
}
\author{
George G. Vega Yon
}

