% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.tw_Class_words}
\alias{plot.tw_Class_words}
\title{Plot words}
\usage{
\method{plot}{tw_Class_words}(x, y = NULL, scale = c(4, 0.5),
  min.freq = 100, max.n.words = 100, around = NULL,
  colors = brewer.pal(min(c(9, max.n.words)), "Blues"), ...)
}
\description{
Plot words
}

