% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{senate_tweets}
\alias{senate_tweets}
\title{US senate tweets}
\format{A data.frame with 19,786 rows and 14 variables}
\source{
\url{http://dev.twitter.com}
}
\description{
This dataset contains tweets from the US senate downloaded using the 'twitter'
API tool. It contains the following columns
}
\details{
\itemize{
 \item screen_name
 \item in_reply_to_screen_name
 \item user_id
 \item created_at
 \item id
 \item text
 \item source
 \item truncated
 \item retweet_count
 \item favorite_count
 \item favorited
 \item retweeted
 \item coordinates
 \item source_name
}
}

