% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_followers_list}
\alias{tw_api_get_followers_list}
\title{Get Followers list}
\usage{
tw_api_get_followers_list(screen_name = NULL, twitter_token, user_id = NULL,
  cursor = NULL, count = 200, skip_status = NULL,
  include_user_entities = "false", max.n = NULL, ...)
}
\arguments{
\item{screen_name}{Screen name}

\item{twitter_token}{An object of class \link[httr:oauth1.0_token]{Token1.0} as
generated by \link{tw_gen_token}.}

\item{user_id}{User id}

\item{cursor}{(advanced)}

\item{count}{Number of registries per page (advanced)}

\item{skip_status}{It self}

\item{include_user_entities}{(advanced)}

\item{max.n}{Number of followers to download}

\item{...}{Further arguments to be passed to \code{\link[=GET]{GET()}}}
}
\value{
A list
}
\description{
Get Followers list
}
\details{
\subsection{From Twitter}{Returns a cursored collection of user
objects for users following the specified user.

At this time, results are ordered with the most recent following first --
however, this ordering is subject to unannounced change and eventual
consistency issues.}
}
\references{
Twitter REST API (GET followers/list)
https://dev.twitter.com/rest/reference/get/followers/list
}
\seealso{
Other API functions: \code{\link{tw_api_get_followers_ids}},
  \code{\link{tw_api_get_friends_ids}},
  \code{\link{tw_api_get_search_tweets}},
  \code{\link{tw_api_get_statuses_sample}},
  \code{\link{tw_api_get_statuses_user_timeline}},
  \code{\link{tw_api_get_trends_place}},
  \code{\link{tw_api_get_users_search}},
  \code{\link{tw_api_get_users_show}},
  \code{\link{tw_api_trends_available}},
  \code{\link{tw_gen_token}}
}
\concept{API functions}
