% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_friends_ids}
\alias{tw_api_get_friends_ids}
\title{GET friends/ids}
\usage{
tw_api_get_friends_ids(screen_name = NULL, twitter_token, user_id = NULL,
  cursor = NULL, count = 1000, ...)
}
\arguments{
\item{twitter_token}{}

\item{user_id}{}

\item{cursor}{}

\item{count}{}

\item{...}{}

\item{scree_name}{}
}
\value{
A vector
}
\description{
GET friends/ids
}

