% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_search_tweets}
\alias{tw_api_get_search_tweets}
\title{Gets status updates (tweets) via serach}
\usage{
tw_api_get_search_tweets(q, twitter_token, geocode = NULL, lang = NULL,
  locale = NULL, result_type = c("mixed", "recent", "popular"),
  count = 100, until = "9999-99-99", since_id = NULL, max_id = NULL,
  quietly = FALSE, ...)
}
\arguments{
\item{q}{Search query}

\item{twitter_token}{Token}

\item{geocode}{character with 'latitude,longitude,radius', where radius units must be specified as either 'mi' (miles) or 'km' (kilometers)}

\item{lang}{Restricts language}

\item{locale}{Specifies the locale of the query}

\item{result_type}{Either of the options especified}

\item{count}{Number of statuses to get}

\item{until}{character specifying the limit date (tweets before than) as YYYY-MM-DD}

\item{since_id}{Returns results with an ID greater than (that is, more recent than) the specified ID}

\item{max_id}{Returns results with an ID less than (that is, older than) or equal to the specified ID}

\item{quietly}{Whether or not to show the 'success' message}

\item{...}{Additional arguments passed to \code{\link{GET}}}
}
\value{
A data.frame with tweets (if success), with the following columns: 
\itemize{
\item \code{result_type}
\item \code{screen_name}
\item \code{in_reply_to_screen_name}
\item \code{user_id}
\item \code{created_at}
\item \code{id}
\item \code{text}
\item \code{source}
\item \code{truncated}
\item \code{retweet_count}
\item \code{favorite_count}
\item \code{favorited}
\item \code{retweeted}
\item \code{coordinates}
\item \code{source_name}
}

otherwise returns \code{NULL}.
}
\description{
Using the twitter API, gets the status updates via search
}
\details{
Keep in mind that the search index has a 7-day limit. In other words,
no tweets will be found for a date older than one week.
\subsection{From Twitter}{Returns a collection of relevant Tweets matching a 
specified query.

Please note that Twitter’s search service and, by extension, the Search API
is not meant to be an exhaustive source of Tweets. Not all Tweets will be
indexed or made available via the search interface.}
}
\examples{
\dontrun{
# Getting the twitts (first gen the token)
key <- tw_gen_token('myapp','key', 'secret')

# Making a query
x <- tw_api_get_search_tweets('lovewins', key)
}
}
\author{
George G. Vega Yon
}
\references{
Twitter REST API (GET search/tweets)
\url{https://dev.twitter.com/rest/reference/get/search/tweets}
}
\seealso{
\code{\link{tw_extract}}

Other API functions: \code{\link{tw_api_get_followers_ids}},
  \code{\link{tw_api_get_followers_list}},
  \code{\link{tw_api_get_friends_ids}},
  \code{\link{tw_api_get_statuses_sample}},
  \code{\link{tw_api_get_statuses_user_timeline}},
  \code{\link{tw_api_get_trends_place}},
  \code{\link{tw_api_get_users_search}},
  \code{\link{tw_api_get_users_show}},
  \code{\link{tw_api_trends_available}},
  \code{\link{tw_gen_token}}
}

