% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_statuses_user_timeline}
\alias{tw_api_get_statuses_user_timeline}
\title{Gets status updates (tweets) from a given user}
\usage{
tw_api_get_statuses_user_timeline(screen_name = NULL, twitter_token,
  user_id = NULL, since_id = NULL, count = 100, max_id = NULL,
  exclude_replies = NULL, include_rts = NULL, quietly = FALSE, ...)
}
\arguments{
\item{screen_name}{of the user}

\item{twitter_token}{Token}

\item{user_id}{The ID of the user for whom to return results for}

\item{since_id}{Returns results with an ID greater than (that is, more recent than) the specified ID}

\item{count}{Number of statuses to get}

\item{max_id}{Returns results with an ID less than (that is, older than) or equal to the specified ID}

\item{exclude_replies}{This parameter will prevent replies from appearing in the returned timeline}

\item{include_rts}{When set to false, the timeline will strip any native retweets}

\item{quietly}{Whether or not to show the 'success' message}

\item{...}{Additional arguments passed to \code{\link{GET}}}
}
\value{
A data.frame with tweets (if success), with the following columns: 
\itemize{
\item \code{screen_name}
\item \code{in_reply_to_screen_name}
\item \code{user_id}
\item \code{created_at}
\item \code{id}
\item \code{text}
\item \code{source}
\item \code{truncated}
\item \code{retweet_count}
\item \code{favourites_count}
\item \code{favorited}
\item \code{retweeted}
\item \code{coordinates}
\item \code{source_name}
}

otherwise returns \code{NULL}.
}
\description{
Using the twitter API, gets the status updates of a given user (up to 3,200)
}
\details{
This function is designed to be applied to a large list of twitter
accounts, see the example below.
\subsection{From twitter}{Returns a collection of the most recent Tweets
posted by the user indicated by the screen_name or user_id parameters.
}
}
\examples{
\dontrun{
# List of twitter accounts
users <- c('MarsRovers', 'senatormenendez', 'sciencemagazine')

# Getting the twitts (first gen the token)
key <- tw_gen_token('myapp','key', 'secret')
tweets <- lapply(users, tw_api_get_statuses_user_timeline, twitter_token=key)

# Processing the data (and taking a look)
tweets <- do.call(rbind, tweets)
head(tweets)
}
}
\references{
Twitter REST API (GET statuses/user_timeline)
\url{https://dev.twitter.com/rest/reference/get/statuses/user_timeline}
}
\seealso{
\code{\link{tw_extract}}

Other API functions: \code{\link{tw_api_get_followers_ids}},
  \code{\link{tw_api_get_followers_list}},
  \code{\link{tw_api_get_friends_ids}},
  \code{\link{tw_api_get_search_tweets}},
  \code{\link{tw_api_get_statuses_sample}},
  \code{\link{tw_api_get_trends_place}},
  \code{\link{tw_api_get_users_search}},
  \code{\link{tw_api_get_users_show}},
  \code{\link{tw_api_trends_available}},
  \code{\link{tw_gen_token}}
}
\author{
George G. Vega Yon
}
