% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_timeline}
\alias{tw_api_get_timeline}
\title{Gets status updates (tweets) from a given user}
\usage{
tw_api_get_timeline(usr, count = 100, ...)
}
\arguments{
\item{usr}{screen_name of the user}

\item{count}{Number of statuses to get}
}
\value{
A data.frame with tweets
}
\description{
Gets status updates (tweets) from a given user
}

