% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_trends_place}
\alias{tw_api_get_trends_place}
\title{Gets the trends for a given place}
\usage{
tw_api_get_trends_place(id, exclude = FALSE, ...)
}
\arguments{
\item{id}{Id of the place}
}
\description{
Gets the trends for a given place
}

