% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_users_show}
\alias{tw_api_get_users_show}
\title{Get user information}
\usage{
tw_api_get_users_show(usr, twitter_token, quietly = FALSE, ...)
}
\arguments{
\item{usr}{User screen name}

\item{twitter_token}{Token}

\item{quietly}{Whether to show the 'success' message or not}

\item{...}{Additional arguments passed to \code{\link{GET}}}
}
\value{
A data.frame with info of the usr.
\itemize{
\item \code{id}
\item \code{name}
\item \code{screen_name}
\item \code{contributors_enabled}
\item \code{created_at}
\item \code{default_profile}
\item \code{default_profile_image}
\item \code{description}
\item \code{favourites_count}
\item \code{followers_count}
\item \code{friends_count}
\item \code{geo_enabled}
\item \code{is_translator}
\item \code{lang}
\item \code{listed_count}
\item \code{location}
\item \code{profile_image_url}
\item \code{profile_image_url_https}
\item \code{protected}
\item \code{statuses_count}
\item \code{time_zone}
\item \code{utc_offset}
\item \code{verified}
}
}
\description{
Get user information
}
\details{
Using the twitter api, get information about a twitter account
}
\examples{
\dontrun{
tw_api_get_users_show('gvegayon')
}
}
\references{
Twitter REST API (GET users/show)
\url{https://dev.twitter.com/rest/reference/get/users/show}
}

