% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_wait}
\alias{tw_api_wait}
\title{Wait in minutes to reconnect to the API}
\usage{
tw_api_wait(minutes = 1)
}
\arguments{
\item{minutes}{Number of minutes to wait}
}
\value{
Void
}
\description{
Wait in minutes to reconnect to the API
}
\details{
Internal use only (it's just a timer...)
}

