% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text_processing.R
\name{tw_extract}
\alias{tw_extract}
\title{Extract info from tweets}
\usage{
tw_extract(txt, obj = c("email", "mention", "hashtag", "url"),
  normalize = TRUE)
}
\arguments{
\item{txt}{Character}

\item{obj}{List of objects to extract}

\item{normalize}{bool whether or not to normalize emails, hashtags and mentions (to lower)}
}
\value{
List
}
\description{
Extract info from tweets
}
\examples{
\dontrun{
head(tw_extract(tweets$text))
}
}
\author{
George G. Vega Yon
}

