% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_gen_token}
\alias{tw_gen_token}
\title{Generate twitter tokens}
\usage{
tw_gen_token(appname, key, secret = NULL)
}
\arguments{
\item{appname}{Character scalar. The name of your twitter app.}

\item{key}{Character scalar. Your API key provided by twitter.}

\item{secret}{Character scalar. Your API secret provided by twitter.}
}
\value{
A Token1.0 reference class (RC) object.
}
\description{
Generate twitter tokens
}
\details{
This function is a wrapper of \code{\link[httr:oauth_app]{httr::oauth_app()}}. To be able to use most of
twitter's API service you need to sign-in with your twitter account at
https://apps.twitter.com/]
}
\examples{
\dontrun{
  tokens <- tw_gen_token('mytwitterapp','xxxxxxxx','yyyyyyyy')
  tw_api_get_usr_profile('gvegayon',tokens)
}
}
\seealso{
Other API functions: \code{\link{tw_api_get_followers_ids}},
  \code{\link{tw_api_get_followers_list}},
  \code{\link{tw_api_get_friends_ids}},
  \code{\link{tw_api_get_search_tweets}},
  \code{\link{tw_api_get_statuses_sample}},
  \code{\link{tw_api_get_statuses_user_timeline}},
  \code{\link{tw_api_get_trends_place}},
  \code{\link{tw_api_get_users_search}},
  \code{\link{tw_api_get_users_show}},
  \code{\link{tw_api_trends_available}}
}
\concept{API functions}
