% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gender.R
\name{tw_gender}
\alias{tw_gender}
\title{Matches text with male/female names}
\usage{
tw_gender(x, male = NULL, female = NULL, lan = c("es", "en"),
  rmNoAlpha = TRUE)
}
\arguments{
\item{x}{Character vector of names/text to analyze and match}

\item{male}{Character vector of male names}

\item{female}{Character vector of female names}

\item{lan}{Languaje of the names}

\item{rmNoAlpha}{Whether or not to remove no alpha characters}
}
\description{
Matches text with male/female names
}
\details{
The char match is written in C++, which is why it should be fast
}
\section{Names datasets}{

}
\references{
For english names OpenGenderTracking project
\url{https://github.com/OpenGenderTracking/globalnamedata}

For spanish names
\url{https://github.com/apalancat/spanish-names}
}

