% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_crawling.R
\name{tw_get_tw_account}
\alias{tw_get_tw_account}
\title{Retrieve twitter accounts from a website}
\usage{
tw_get_tw_account(uri, redirect = TRUE, normalize = TRUE, quiet = TRUE)
}
\arguments{
\item{uri}{Character scalar. Web address of the website from which to extract
twitter accounts.}

\item{redirect}{Whether to follow up the link or not (redirection)}

\item{normalize}{When \code{TRUE} returns the accounts as lower-case.}

\item{quiet}{When \code{FALSE} suppress the message}
}
\value{
A character vector with twitter accounts' screen name found if any.
}
\description{
Retrieve twitter accounts from a website
}
\examples{

\dontrun{
# Twitter accounts from US gov officials (out dated)
tw_get_tw_account('http://twitter.pbworks.com/w/page/1779986/USGovernment')

# Caltech
tw_get_tw_account('http://caltech.edu')
}

}
\seealso{
Other text processors: \code{\link{tw_extract}},
  \code{\link{tw_words}}
}
\author{
George G. Vega Yon
}
\concept{text processors}
