% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{tw_leaflet}
\alias{tw_leaflet}
\title{Creates a leaflet map}
\usage{
tw_leaflet(dat, coordinates = NULL, popup = NULL, lat = NULL,
  lng = NULL, radii = ~sqrt(n) * 5000, weight = 1,
  cluster.method = "centroid", nclusters = min(c(50, nrow(dat)\%/\%3)), ...)
}
\arguments{
\item{dat}{A twitter dataset (with coordinates)}

\item{coordinates}{Name of the \code{coordinates} variable (see details)}

\item{popup}{Name of the grouping variable (for example, \code{screen_name})}

\item{lat}{Name of the latitude variable}

\item{lng}{Name of the longitude variable}

\item{radii}{Formula to compute the radious of the circles}

\item{weight}{Thickness of the circles' borders}

\item{cluster.method}{Clustering method (see \code{\link{hclust}})}

\item{nclusters}{Max number of clusters to include}

\item{...}{Further arguments to be passed to \code{\link{addCircles}}}
}
\value{
A map object (see the \code{\link{leaflet}} package)
}
\description{
This function works as a wrapper of the \code{leaflet} function, creating a
map with circles which sizes are set by the number of observations in those
coordinates.
}
\details{
The \code{coordinates} must be in the format of longitude:latitude
(as the twitter API returns). 

In order to improve visualization, the function performs Hierarchical
Clustering via \code{\link{hclust}} (from the stats package), grouping
observations by geo coordinates. For each cluster, the final lat/lng coords
are defined as the mean within the cluster.

In the case of \code{radii}, it is measured in meters (in the map). The
\code{n} is the number of observations in that cluster.

For performance considerations, it is recommended not to use more than
1,000 observations (Try using a random sample from your data!) as 
\code{hclust} requires computing square matrices.
}
\examples{
\dontrun{
# Getting the data
data(senate_tweets)

# using formulas to pass the variables names
tw_leaflet(senate_tweets,~coordinates,~screen_name)
tw_leaflet(senate_tweets,~coordinates)

# Using characters to pass the variables names
tw_leaflet(senate_tweets,'coordinates','screen_name')
tw_leaflet(senate_tweets,'coordinates')

# Aggregating until get only 3 big groups
tw_leaflet(senate_tweets,'coordinates', nclusters=3)
}
}
\author{
George G. Vega Yon
}
