% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sentiment.R
\name{tw_sentiment}
\alias{tw_sentiment}
\title{Computes sentiment score}
\usage{
tw_sentiment(text, pos = NULL, neg = NULL, lang = "en")
}
\arguments{
\item{text}{Character vector}

\item{pos}{Positive words}

\item{neg}{Negative words}

\item{lang}{Languaje}
}
\value{
Numeric Vector with scores
}
\description{
Computes sentiment score by counting the number of positive and negative terms.
}
\details{
By default uses an english lexicon downloaded from
\url{http://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html}.
}
\seealso{
\code{\link{sentiment_lexicon_pos_en}} \code{\link{sentiment_lexicon_neg_en}}
}

