% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/statistics.R
\name{plot.tw_Class_table}
\alias{plot.tw_Class_table}
\alias{tw_table}
\alias{tw_Class_table}
\alias{tw_table.tw_Class_extract}
\title{Creates a table of frequencies}
\usage{
\method{plot}{tw_Class_table}(x, y = NULL, nentries = 20,
  caption = "Most popular hashtags", options = list(pageLength = 5), ...)

tw_table(x, ...)

\method{tw_table}{tw_Class_extract}(x, obj = c("email", "mention", "hashtag",
  "url"), ...)
}
\arguments{
\item{x}{Either an object of class \link{tw_Class_extract}, or \link{tw_Class_table}.}

\item{y}{Ignored}

\item{nentries}{Number of rows to include in the table}

\item{caption, options}{Passed to \code{\link[DT:datatable]{DT::datatable()}}.}

\item{...}{Further arguments to be passed to the method}

\item{obj}{A character indicating}
}
\description{
See \code{\link[=tw_extract]{tw_extract()}}
}
\details{
In the case of \code{plot.tw_Class_table}, \code{...} are passed to \code{\link[DT:datatable]{DT::datatable()}}
}
\section{Methods (by class)}{
\itemize{
\item \code{tw_Class_extract}: Makes a table out of the output of \code{\link[=tw_extract]{tw_extract()}}
}}

\examples{
# Loading tweets
data(senate_tweets)
head(senate_tweets$text)

# Extracting elements and creating tables
x <- tw_extract(senate_tweets$text)

head(tw_table(x,'mention'))
}
\seealso{
Other statistics: \code{\link{plot.tw_Class_ts}},
  \code{\link{tw_sentiment}}
}
\author{
George G. Vega Yon
}
\concept{statistics}
