% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/statistics.R
\name{plot.tw_Class_ts}
\alias{plot.tw_Class_ts}
\alias{tw_timeseries}
\alias{tw_Class_ts}
\title{Creates a time series object with frecuencies}
\usage{
\method{plot}{tw_Class_ts}(x, rangeSelector = TRUE,
  main = "Number of daily tweets", xlab = NULL, ylab = NULL,
  group = NULL, width = 600, height = 400, ...)

tw_timeseries(obj, created_at, units = c("secs", "mins", "hours", "days"),
  nseries = 5, span = 30)
}
\arguments{
\item{x}{An object of class \link{tw_Class_ts}.}

\item{rangeSelector}{Logical scalar. Wheather or not to include a time range
selector.}

\item{main, xlab, ylab, group, width, height}{Passed to \code{\link[dygraphs:dygraph]{dygraphs::dygraph()}}.}

\item{...}{Further arguments to be passed to the \code{\link[dygraphs:dygraph]{dygraphs::dygraph()}}.}

\item{obj}{Any of the members of the output list from tw_extract()}

\item{created_at}{\link{POSIXct} class vector}

\item{units}{Time unit}

\item{nseries}{Number of series to include}

\item{span}{Number of time units to consider}
}
\value{
xts class
}
\description{
Creates a time series object with frecuencies
}
\seealso{
Other statistics: \code{\link{plot.tw_Class_table}},
  \code{\link{tw_sentiment}}
}
\author{
George G. Vega Yon

George G. Vega Yon
}
\concept{statistics}
