% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text_processing.R
\name{tw_words}
\alias{tw_words}
\title{Process txt extracting relevant words, i.e. rm stop words and
 symbols.}
\usage{
tw_words(txt, stopw = stopwords("en"), cleanfun = NULL)
}
\arguments{
\item{txt}{A character vector with text to analyze}

\item{stopw}{Set of stop words}

\item{cleanfun}{A function to pass to lapply}
}
\description{
Process txt extracting relevant words, i.e. rm stop words and
 symbols.
}
\author{
George G. Vega Yon
}

