% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_write_json_network}
\alias{tw_write_json_network}
\title{Writes a JSON graph to be used with d3js}
\usage{
tw_write_json_network(graph)
}
\arguments{
\item{graph}{}
}
\description{
Writes a JSON graph to be used with d3js
}

