% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{twitterreport}
\alias{twitterreport}
\alias{twitterreport-package}
\title{twitterreport}
\description{
Out-of-the-Box Analysis and Reporting Tools for Twitter
}
\author{
George G. Vega Yon
}

