% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{warriner_et_al_en}
\alias{warriner_et_al_en}
\title{Warriner et al. Lexicon (en)}
\format{A data.frame with scores (Valence)}
\source{
\url{http://crr.ugent.be/archives/1003}
}
\description{
Warriner et al. Lexicon (en)
}
\details{
Spanish version translated by Daniel Gayo-Avello based on Warriner, et al.
}
\seealso{
Other example datasets: \code{\link{names_female_en}},
  \code{\link{names_female_es}},
  \code{\link{names_male_en}}, \code{\link{names_male_es}},
  \code{\link{senate_tweets}},
  \code{\link{senators_profile}}, \code{\link{senators}},
  \code{\link{sentiment_lexicon_neg_en}},
  \code{\link{sentiment_lexicon_pos_en}},
  \code{\link{warriner_et_al_es}}

Other lexicon datasets: \code{\link{sentiment_lexicon_neg_en}},
  \code{\link{sentiment_lexicon_pos_en}},
  \code{\link{warriner_et_al_es}}
}

