% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{warriner_et_al_es}
\alias{warriner_et_al_es}
\title{Warriner et al. Lexicon (es)}
\format{A data.frame with scores (Valence)}
\source{
\url{http://danigayo.info/blog/index.php?entry=entry130117-183114}
}
\description{
Warriner et al. Lexicon (es)
}

