% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_processing.R
\name{words_closeness}
\alias{words_closeness}
\title{Retrieves a set of words related to a particular word}
\usage{
words_closeness(word, jaccard, criter = 0.01, exact = FALSE)
}
\arguments{
\item{word}{A character word to analyze}

\item{jaccard}{A \code{tw_Class_jaccard} class object}

\item{criter}{Minimun number in the index to show}

\item{exact}{When \code{FALSE} shearch via regular expression}
}
\value{
A data frame containing the list of words that are related
to the specified \code{word}
}
\description{
Retrieves a set of words related to a particular word
}
\details{
After applying the \code{\link{jaccard_coef}} function, the 
resulting object can be analyzed with this function.
}
\examples{
\dontrun{
# Computing the jaccard coefficient
jaccard <- jaccard_coef(tweets$text)

# See what words are related with abortion
words_closeness('abortion',jaccard,.001)
}
}
