% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{names_female_en}
\alias{names_female_en}
\title{English female names}
\format{A character vector with 80,148 elements}
\source{
OpenGenderTracking project \url{https://github.com/OpenGenderTracking/globalnamedata}
}
\description{
English female names
}
\seealso{
Other example datasets: \code{\link{names_female_es}},
  \code{\link{names_male_en}}, \code{\link{names_male_es}},
  \code{\link{senate_tweets}},
  \code{\link{senators_profile}}, \code{\link{senators}},
  \code{\link{sentiment_lexicon_neg_en}},
  \code{\link{sentiment_lexicon_pos_en}},
  \code{\link{warriner_et_al_en}},
  \code{\link{warriner_et_al_es}}

Other names datasets: \code{\link{names_female_es}},
  \code{\link{names_male_en}}, \code{\link{names_male_es}}
}

