% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{names_male_es}
\alias{names_male_es}
\title{Spanish male names}
\format{A character vector with 25,442 elements}
\source{
\url{https://github.com/apalancat/spanish-names}
}
\description{
Spanish male names
}
\seealso{
Other example datasets: \code{\link{names_female_en}},
  \code{\link{names_female_es}},
  \code{\link{names_male_en}}, \code{\link{senate_tweets}},
  \code{\link{senators_profile}}, \code{\link{senators}},
  \code{\link{sentiment_lexicon_neg_en}},
  \code{\link{sentiment_lexicon_pos_en}},
  \code{\link{warriner_et_al_en}},
  \code{\link{warriner_et_al_es}}

Other names datasets: \code{\link{names_female_en}},
  \code{\link{names_female_es}},
  \code{\link{names_male_en}}
}
