% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.tw_Class_graph}
\alias{plot.tw_Class_graph}
\title{Plots a network using D3js}
\usage{
\method{plot}{tw_Class_graph}(x, y = NULL, nodelabel = "name",
  opacity = 0.9, opacityNoHover = 0.5, legend = TRUE, ...)
}
\arguments{
\item{x}{A tw_Class_graph object (see \code{\link{tw_network}})}

\item{y}{Ignored}

\item{nodelabel}{Name of the column that containts the name}

\item{opacity}{See \code{\link{forceNetwork}}}

\item{opacityNoHover}{Idem}

\item{legend}{Whether to include a legend or not}

\item{...}{Further arguments to pass to forceNetwork}
}
\value{
A D3js graph
}
\description{
Plots a network using D3js
}
