% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{senate_tweets}
\alias{senate_tweets}
\title{US senate tweets}
\format{A data.frame with 19,786 rows and 14 variables}
\source{
\url{http://dev.twitter.com}
}
\description{
This dataset contains tweets from the US senate downloaded using the 'twitter'
API tool. It contains the following columns
}
\details{
\itemize{
 \item \code{screen_name}
 \item \code{in_reply_to_screen_name}
 \item \code{user_id}
 \item \code{created_at}
 \item \code{id}
 \item \code{text}
 \item \code{source}
 \item \code{truncated}
 \item \code{retweet_count}
 \item \code{favorite_count}
 \item \code{favorited}
 \item \code{retweeted}
 \item \code{coordinates}
 \item \code{source_name}
}
}

