% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{senate_tweets}
\alias{senate_tweets}
\title{US senate tweets}
\format{A data.frame with 19,786 rows and 14 variables}
\source{
http://dev.twitter.com
}
\description{
This dataset contains tweets from the US senate downloaded using the 'twitter'
API tool. It contains the following columns
}
\details{
\itemize{
\item \code{screen_name}
\item \code{in_reply_to_screen_name}
\item \code{user_id}
\item \code{created_at}
\item \code{id}
\item \code{text}
\item \code{source}
\item \code{truncated}
\item \code{retweet_count}
\item \code{favorite_count}
\item \code{favorited}
\item \code{retweeted}
\item \code{coordinates}
\item \code{source_name}
}
}
\seealso{
Other example datasets: \code{\link{names_female_en}},
  \code{\link{names_female_es}},
  \code{\link{names_male_en}}, \code{\link{names_male_es}},
  \code{\link{senators_profile}}, \code{\link{senators}},
  \code{\link{sentiment_lexicon_neg_en}},
  \code{\link{sentiment_lexicon_pos_en}},
  \code{\link{warriner_et_al_en}},
  \code{\link{warriner_et_al_es}}
}
\concept{example datasets}
