% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{senators}
\alias{senators}
\title{US senators info from senate.gov}
\format{A data.frame with 100 rows and 7 variables}
\source{
\url{http://www.senate.gov/senators/contact/}
}
\description{
Extracted using webcrawling
}
\details{
Name
\itemize{
\item party
\item State
\item Addr
\item phone
\item website
\item class
}
}

