% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{senators}
\alias{senators}
\title{US senators info from senate.gov}
\format{A data.frame with 100 rows and 7 variables}
\source{
\url{http://www.senate.gov/senators/contact/}
}
\description{
Extracted using webcrawling, the dataset contains the following columns:
}
\details{
\itemize{
 \item \code{Name}
 \item \code{party}
 \item \code{State}
 \item \code{Addr}
 \item \code{phone}
 \item \code{website}
 \item \code{class}
}
}
\seealso{
Other example datasets: \code{\link{names_female_en}},
  \code{\link{names_female_es}},
  \code{\link{names_male_en}}, \code{\link{names_male_es}},
  \code{\link{senate_tweets}},
  \code{\link{senators_profile}},
  \code{\link{sentiment_lexicon_neg_en}},
  \code{\link{sentiment_lexicon_pos_en}},
  \code{\link{warriner_et_al_en}},
  \code{\link{warriner_et_al_es}}
}

