% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{senators_profile}
\alias{senators_profile}
\title{US senators twitter profile}
\format{A data.frame with 100 rows and 23 columns}
\source{
\url{https://dev.twitter.com}
}
\description{
US senators twitter profile
}

