% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sentiment_lexicon_neg_en}
\alias{sentiment_lexicon_neg_en}
\title{Negative Lexicon (en)}
\format{A Character vector}
\source{
\url{http://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html#datasets}
}
\description{
Negative Lexicon (en)
}
\seealso{
Other example datasets: \code{\link{names_female_en}},
  \code{\link{names_female_es}},
  \code{\link{names_male_en}}, \code{\link{names_male_es}},
  \code{\link{senate_tweets}},
  \code{\link{senators_profile}}, \code{\link{senators}},
  \code{\link{sentiment_lexicon_pos_en}},
  \code{\link{warriner_et_al_en}},
  \code{\link{warriner_et_al_es}}

Other lexicon datasets: \code{\link{sentiment_lexicon_pos_en}},
  \code{\link{warriner_et_al_en}},
  \code{\link{warriner_et_al_es}}
}

