% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{sentiment_lexicon_pos_en}
\alias{sentiment_lexicon_pos_en}
\title{Positive Lexicon (en)}
\format{A Character vector}
\source{
\url{http://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html#datasets}
}
\description{
Positive Lexicon (en)
}

