% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_followers_ids}
\alias{tw_api_get_followers_ids}
\title{GET followers/ids}
\usage{
tw_api_get_followers_ids(screen_name = NULL, twitter_token, user_id = NULL,
  cursor = NULL, count = 1000, ...)
}
\arguments{
\item{screen_name}{User screen name}

\item{twitter_token}{Token}

\item{user_id}{User Id number}

\item{cursor}{Page number (cursor)}

\item{count}{Number of ids by page}

\item{...}{Further arguments to be passed to \code{\link{GET}}}
}
\value{
A vector
}
\description{
GET followers/ids
}

