% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_followers_ids}
\alias{tw_api_get_followers_ids}
\title{GET followers/ids}
\usage{
tw_api_get_followers_ids(screen_name = NULL, twitter_token, user_id = NULL,
  cursor = NULL, count = 1000, ...)
}
\arguments{
\item{screen_name}{User screen name}

\item{twitter_token}{An object of class \link[httr:oauth1.0_token]{Token1.0} as
generated by \link{tw_gen_token}.}

\item{user_id}{User Id number}

\item{cursor}{Page number (cursor)}

\item{count}{Number of ids by page}

\item{...}{Further arguments to be passed to \code{\link[=GET]{GET()}}}
}
\value{
A vector
}
\description{
Returns a list of followers IDs of a given user.
}
\details{
\subsection{From twitter}{Returns a cursored collection of user IDs
for every user following the specified user.

At this time, results are ordered with the most recent following first --
however, this ordering is subject to unannounced change and eventual
consistency issues.}
}
\references{
Twitter REST API (GET followers/ids)
https://dev.twitter.com/rest/reference/get/followers/ids
}
\seealso{
Other API functions: \code{\link{tw_api_get_followers_list}},
  \code{\link{tw_api_get_friends_ids}},
  \code{\link{tw_api_get_search_tweets}},
  \code{\link{tw_api_get_statuses_sample}},
  \code{\link{tw_api_get_statuses_user_timeline}},
  \code{\link{tw_api_get_trends_place}},
  \code{\link{tw_api_get_users_search}},
  \code{\link{tw_api_get_users_show}},
  \code{\link{tw_api_trends_available}},
  \code{\link{tw_gen_token}}
}
\concept{API functions}
