% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_followers_list}
\alias{tw_api_get_followers_list}
\title{Get Followers list}
\usage{
tw_api_get_followers_list(screen_name = NULL, twitter_token, user_id = NULL,
  cursor = NULL, count = 200, skip_status = NULL,
  include_user_entities = "false", howmany = NULL, ...)
}
\arguments{
\item{screen_name}{Screen name}

\item{twitter_token}{Twitter token}

\item{user_id}{User id}

\item{cursor}{(advanced)}

\item{count}{Number of registries per page (advanced)}

\item{skip_status}{It self}

\item{include_user_entities}{(advanced)}

\item{howmany}{Number of followers to download}

\item{...}{Further arguments to be passed to \code{\link{GET}}}
}
\value{
A list
}
\description{
Get Followers list
}

