% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_friends_ids}
\alias{tw_api_get_friends_ids}
\title{GET friends/ids}
\usage{
tw_api_get_friends_ids(screen_name = NULL, twitter_token, user_id = NULL,
  cursor = NULL, count = 1000, ...)
}
\arguments{
\item{screen_name}{Screen name}

\item{twitter_token}{An object of class \link[httr:oauth1.0_token]{Token1.0} as
generated by \link{tw_gen_token}.}

\item{user_id}{User ID number}

\item{cursor}{Cursor page}

\item{count}{Number of elements to download}

\item{...}{Further arguments to be passed to \code{\link[=GET]{GET()}}}
}
\value{
A vector
}
\description{
Returns a list of friends of a given user (list of who he/she follows)
}
\details{
\subsection{From Twitter}{Returns a cursored collection of user IDs
for every user the specified user is following (otherwise known as their
``friends'').}
}
\references{
Twitter REST API (GET search/tweets)
https://dev.twitter.com/rest/reference/get/friends/ids
}
\seealso{
Other API functions: \code{\link{tw_api_get_followers_ids}},
  \code{\link{tw_api_get_followers_list}},
  \code{\link{tw_api_get_search_tweets}},
  \code{\link{tw_api_get_statuses_sample}},
  \code{\link{tw_api_get_statuses_user_timeline}},
  \code{\link{tw_api_get_trends_place}},
  \code{\link{tw_api_get_users_search}},
  \code{\link{tw_api_get_users_show}},
  \code{\link{tw_api_trends_available}},
  \code{\link{tw_gen_token}}
}
\concept{API functions}
