% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_friends_ids}
\alias{tw_api_get_friends_ids}
\title{GET friends/ids}
\usage{
tw_api_get_friends_ids(screen_name = NULL, twitter_token, user_id = NULL,
  cursor = NULL, count = 1000, ...)
}
\arguments{
\item{screen_name}{Screen name}

\item{twitter_token}{Token}

\item{user_id}{User ID number}

\item{cursor}{Cursor page}

\item{count}{Number of elements to download}

\item{...}{Further arguments to be passed to \code{\link{GET}}}
}
\value{
A vector
}
\description{
GET friends/ids
}

