% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_statuses_sample}
\alias{tw_api_get_statuses_sample}
\title{Gets a sample through the sample API}
\usage{
tw_api_get_statuses_sample(twitter_token, Timeout = 60, ...)
}
\arguments{
\item{twitter_token}{Token}

\item{Timeout}{Number of seconds}

\item{...}{Additional parameters to be passed to \code{\link{GET}}}
}
\description{
Gets a sample through the sample API
}

