% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_statuses_sample}
\alias{tw_api_get_statuses_sample}
\title{Gets a sample through the sample API}
\usage{
tw_api_get_statuses_sample(twitter_token, Timeout = 60, ...)
}
\arguments{
\item{twitter_token}{Token}

\item{Timeout}{Number of seconds}

\item{...}{Additional parameters to be passed to \code{\link{GET}}}
}
\description{
ON DEVELOPMENT
}
\details{
\subsection{From Twitter}{Returns a small random sample of all
public statuses. The Tweets returned by the default access level are the
same, so if two different clients connect to this endpoint, they will see the
same Tweets.}
}
\references{
Twitter Streaming API (GET statuses/sample)
\url{https://dev.twitter.com/streaming/reference/get/statuses/sample}
}
\seealso{
Other API functions: \code{\link{tw_api_get_followers_ids}},
  \code{\link{tw_api_get_followers_list}},
  \code{\link{tw_api_get_friends_ids}},
  \code{\link{tw_api_get_search_tweets}},
  \code{\link{tw_api_get_statuses_user_timeline}},
  \code{\link{tw_api_get_trends_place}},
  \code{\link{tw_api_get_users_search}},
  \code{\link{tw_api_get_users_show}},
  \code{\link{tw_api_trends_available}},
  \code{\link{tw_gen_token}}
}

