% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_timeline}
\alias{tw_api_get_timeline}
\title{Gets status updates (tweets) from a given user}
\usage{
tw_api_get_timeline(usr, twitter_token, count = 100, quietly = FALSE, ...)
}
\arguments{
\item{usr}{screen_name of the user}

\item{twitter_token}{Token}

\item{count}{Number of statuses to get}

\item{quietly}{Whether or not to show the 'success' message}

\item{...}{Additional arguments passed to \code{\link{GET}}}
}
\value{
A data.frame with tweets
}
\description{
Gets status updates (tweets) from a given user
}

