% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_trends_place}
\alias{tw_api_get_trends_place}
\title{Gets the trends for a given place}
\usage{
tw_api_get_trends_place(id, twitter_token, exclude = FALSE, ...)
}
\arguments{
\item{id}{Id of the place}

\item{twitter_token}{Token}

\item{exclude}{Whether to exclude hashtags or not}

\item{...}{Ignored}
}
\description{
Gets the trends for a given place
}
\references{
Twitter REST API (GET trends/place)
\url{https://dev.twitter.com/rest/reference/get/trends/place}
}

