% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_trends_place}
\alias{tw_api_get_trends_place}
\title{Gets the trends for a given place}
\usage{
tw_api_get_trends_place(id, twitter_token, exclude = FALSE, ...)
}
\arguments{
\item{id}{Id of the place}

\item{twitter_token}{Token}

\item{exclude}{Whether to exclude hashtags or not}

\item{...}{Ignored}
}
\description{
Uses the output from \code{\link{tw_api_trends_available}}
}
\details{
\subsection{From Twitter}{Returns the top 10 trending topics for a
specific WOEID, if trending information is available for it.

The response is an array of ``trend'' objects that encode the name of the
trending topic, the query parameter that can be used to search for the topic
on \href{http://search.twitter.com/}{Twitter Search}, and the Twitter Search
URL.

This information is cached for 5 minutes. Requesting more frequently than
that will not return any more data, and will count against your rate limit
usage.}
}
\references{
Twitter REST API (GET trends/place)
\url{https://dev.twitter.com/rest/reference/get/trends/place}
}
\seealso{
Other API functions: \code{\link{tw_api_get_followers_ids}},
  \code{\link{tw_api_get_followers_list}},
  \code{\link{tw_api_get_friends_ids}},
  \code{\link{tw_api_get_search_tweets}},
  \code{\link{tw_api_get_statuses_sample}},
  \code{\link{tw_api_get_statuses_user_timeline}},
  \code{\link{tw_api_get_users_search}},
  \code{\link{tw_api_get_users_show}},
  \code{\link{tw_api_trends_available}},
  \code{\link{tw_gen_token}}
}
