% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_users_search}
\alias{tw_api_get_users_search}
\title{Search users}
\usage{
tw_api_get_users_search(q, twitter_token, page = NULL, count = 20, ...)
}
\arguments{
\item{q}{Query}

\item{twitter_token}{key}

\item{page}{Page number to retrieve}

\item{count}{Number of accounts per page}

\item{...}{Further parameters to be passed to \code{\link{GET}}}
}
\value{
A list of twitter accounts
}
\description{
Search users via approximate string matching
}
\details{
\subsection{From Twitter}{Provides a simple, relevance-based search interface to public user
accounts on Twitter. Try querying by topical interest, full name, company name,
location, or other criteria. Exact match searches are not supported.}
}
\references{
Twitter REST API (GET users/search) \url{https://dev.twitter.com/rest/reference/get/users/search}
}

