% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_get_usr_profile}
\alias{tw_api_get_usr_profile}
\title{Get user information}
\usage{
tw_api_get_usr_profile(usr, twitter_token, quietly = FALSE, ...)
}
\arguments{
\item{usr}{User screen name}

\item{twitter_token}{Token}

\item{quietly}{Whether to show the 'success' message or not}

\item{...}{Additional arguments passed to \code{\link{GET}}}
}
\value{
A data.frame with info of the usr.
}
\description{
Get user information
}
\details{
Using the twitter api, get information about a twitter account
}
\examples{
\dontrun{
tw_api_get_usr_profile('gvegayon')
}
}

