% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_api_trends_available}
\alias{tw_api_trends_available}
\title{Get codes from the places where trends are availables}
\usage{
tw_api_trends_available(twitter_token, ...)
}
\arguments{
\item{twitter_token}{Token}

\item{...}{Ignored}
}
\description{
Get codes from the places where trends are availables
}
\references{
Twitter REST API (GET trends/available)
\url{https://dev.twitter.com/rest/reference/get/trends/available}
}

