% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_gen_token}
\alias{tw_gen_token}
\title{Generate twitter tokens}
\usage{
tw_gen_token(appname, key, secret = NULL)
}
\arguments{
\item{appname}{See \code{\link{oauth_app}}}

\item{key}{See \code{\link{oauth_app}}.}

\item{secret}{See \code{\link{oauth_app}}}
}
\value{
A Token1.0 reference class (RC) object.
}
\description{
Generate twitter tokens
}
\examples{
\dontrun{
tokens <- tw_gen_token('mytwitterapp','xxxxxxxx','yyyyyyyy')
tw_api_get_usr_profile('gvegayon',tokens)
}
}

