% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_api.R
\name{tw_gen_token}
\alias{tw_gen_token}
\title{Generate twitter tokens}
\usage{
tw_gen_token(appname, key, secret = NULL)
}
\arguments{
\item{appname}{See \code{\link{oauth_app}}}

\item{key}{See \code{\link{oauth_app}}.}

\item{secret}{See \code{\link{oauth_app}}}
}
\value{
A Token1.0 reference class (RC) object.
}
\description{
Generate twitter tokens
}
\examples{
\dontrun{
tokens <- tw_gen_token('mytwitterapp','xxxxxxxx','yyyyyyyy')
tw_api_get_usr_profile('gvegayon',tokens)
}
}
\seealso{
Other API functions: \code{\link{tw_api_get_followers_ids}},
  \code{\link{tw_api_get_followers_list}},
  \code{\link{tw_api_get_friends_ids}},
  \code{\link{tw_api_get_search_tweets}},
  \code{\link{tw_api_get_statuses_sample}},
  \code{\link{tw_api_get_statuses_user_timeline}},
  \code{\link{tw_api_get_trends_place}},
  \code{\link{tw_api_get_users_search}},
  \code{\link{tw_api_get_users_show}},
  \code{\link{tw_api_trends_available}}
}

