% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/web_crawling.R
\name{tw_get_tw_account}
\alias{tw_get_tw_account}
\title{Function to enter the website and get the twitter account}
\usage{
tw_get_tw_account(uri, redirect = TRUE)
}
\arguments{
\item{uri}{Web address of the website from which to extract twitter accounts}

\item{redirect}{Whether to follow up the link or not (redirection)}
}
\description{
Function to enter the website and get the twitter account
}
\examples{
\dontrun{
tw_get_tw_account('http://twitter.pbworks.com/w/page/1779986/USGovernment')
tw_get_tw_account('http://caltech.edu')
}
}
\author{
George G. Vega Yon
}

