% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gender.R
\name{tw_names_db}
\alias{tw_names_db}
\title{Retrieve the database of}
\usage{
tw_names_db(lan, male = TRUE)
}
\arguments{
\item{lan}{Languaje}

\item{male}{TRUE loads male names}
}
\value{
A factor vector indicating gender or unidentified
}
\description{
Retrieve the database of
}
\examples{
# Retrieving female spanish names
fem_es <- tw_names_db('es',FALSE)

# Retrieving male english names
mal_en <- tw_names_db('en')
}

