% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/networks.R
\name{tw_network}
\alias{tw_network}
\title{Creates conversation graph (directed)}
\usage{
tw_network(source, target, onlyFrom = FALSE, excludeSelf = TRUE,
  minInteract = 1, group = NULL)
}
\arguments{
\item{source}{Vector of screen_name}

\item{target}{List of vectors of mentions (output from tw_extract)}

\item{onlyFrom}{Whether to fileter the links to those only where
source and target are in the \code{source} vector}

\item{excludeSelf}{Whether to exclude selflinks}

\item{minInteract}{Minimun number of interactions to consider (links
below this number will be excluded}

\item{group}{Data frame with two columns: name & group}
}
\description{
Creates conversation graph (directed)
}
\author{
George G. Vega Yon
}

