% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{tw_network}
\alias{tw_network}
\title{Creates graph (directed)}
\usage{
tw_network(source, target, only.from = FALSE, exclude.self = TRUE,
  min.interact = 1, group = NULL, size = NULL, ignore.case = TRUE)
}
\arguments{
\item{source}{Vector of \emph{screen_name}}

\item{target}{List of vectors of \emph{mentions} (output from \code{tw_extract})}

\item{only.from}{Whether to filter the links to those only where
source and target are in the \code{source} vector}

\item{exclude.self}{Whether to exclude self-links}

\item{min.interact}{Minimun number of interactions to consider (links
below this number will be excluded)}

\item{group}{Data frame with two columns: name & group}

\item{size}{A data frame with two columns: name & size}

\item{ignore.case}{When \code{TRUE} converts all of \code{source} and \code{target}
to lower-case.}
}
\value{
A two-element list containing two data.frames, nodes and links of
class \code{tw_Class_graph} (to be used with \code{\link[=plot.tw_Class_graph]{plot.tw_Class_graph()}}.
The nodes data.frame includes two columns, \code{id}, \code{name} and
\code{group}. The links data.frame includes three columns, \code{source},
\code{target} and \code{value}.
}
\description{
Using the output of \code{\link[=tw_extract]{tw_extract()}} and the author of the message,
creates a graph
}
\details{
The \code{value} column in the \code{links} dataframe (see \strong{Value})
is computed as the number of connexions between the source and the target.
}
\examples{
\dontrun{
# Loading sample data and retrieving mentions
data(senate_tweets)
mentions <- tw_extract(senate_tweets$text, obj="mention")$mention

# Preparing data for size. Here we are just setting a random size for
# each vertex.
usrs<- tolower(senate_tweets$screen_name)
size <- data.frame(name=unique(usrs),
                   size=exp(runif(length(unique(usrs)))*5))

# Creating the graph
graph <- tw_network(
  usrs, mentions, min.interact = 5, size=size)

# Visualizing the graph
plot(graph)
}
}
\author{
George G. Vega Yon
}
\concept{network functions}
