% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{tw_timeseries}
\alias{tw_timeseries}
\title{Creates a time series object with frecuencies}
\usage{
tw_timeseries(obj, created_at, units = c("secs", "mins", "hours", "days"),
  nseries = 5, span = 30)
}
\arguments{
\item{obj}{Any of the members of the output list from tw_extract()}

\item{created_at}{POSIXct class vector}

\item{units}{Time unit}

\item{nseries}{Number of series to include}

\item{span}{Number of time units to consider}
}
\value{
xts class
}
\description{
Creates a time series object with frecuencies
}
\author{
George G. Vega Yon
}
\seealso{
Other statistics: \code{\link{tw_sentiment}},
  \code{\link{tw_table}}
}

