% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{warriner_et_al_en}
\alias{warriner_et_al_en}
\title{Warriner et al. Lexicon (en)}
\format{A data.frame with scores (Valence)}
\source{
\url{http://crr.ugent.be/archives/1003}
}
\description{
Warriner et al. Lexicon (en)
}
\details{
Spanish version translated by Daniel Gayo-Avello based on Warriner, et al.
}

