% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{warriner_et_al_es}
\alias{warriner_et_al_es}
\title{Warriner et al. Lexicon (es)}
\format{A data.frame with scores (Valence)}
\source{
\url{http://danigayo.info/blog/index.php?entry=entry130117-183114}
}
\description{
Warriner et al. Lexicon (es)
}
\details{
Based on \emph{Warriner, A.B., Kuperman, V., & Brysbaert, M. (2013). Norms of valence, arousal, and dominance for 13,915 English lemmas. Behavior Research Methods, 45, 1191-1207.}
}
\seealso{
Other example datasets: \code{\link{names_female_en}},
  \code{\link{names_female_es}},
  \code{\link{names_male_en}}, \code{\link{names_male_es}},
  \code{\link{senate_tweets}},
  \code{\link{senators_profile}}, \code{\link{senators}},
  \code{\link{sentiment_lexicon_neg_en}},
  \code{\link{sentiment_lexicon_pos_en}},
  \code{\link{warriner_et_al_en}}

Other lexicon datasets: \code{\link{sentiment_lexicon_neg_en}},
  \code{\link{sentiment_lexicon_pos_en}},
  \code{\link{warriner_et_al_en}}
}

